Rendering to screen modified in order to polish unused values and pointers (See description)
Now, the BlitTarget (Object we send to the depthbuffer for each item to print) does not have an SDL_Rect* to resize objects. If it has a determined width and height, it will use its value. If not, the render will see a default -1 and use the original size of the spritesheet.